+++ src/Core/Logistics.Application.Tenant.Contracts/Commands/UpdateLoadCommand.cs
@@ -6,7 +6,7 @@ public sealed class UpdateLoadCommand : RequestBase<DataResult>
     public string? Name { get; set; }
     public string? SourceAddress { get; set; }
     public string? DestinationAddress { get; set; }
-    public decimal DeliveryCost { get; set; }
+    public double DeliveryCost { get; set; }
     public double Distance { get; set; }
     public string? AssignedDispatcherId { get; set; }
     public string? AssignedDriverId { get; set; }
