+++ src/Core/Logistics.Application.Tenant.Contracts/Models/DailyGrossDto.cs
@@ -2,6 +2,6 @@
 
 public record DailyGrossDto(DateTime Date)
 {
-    public decimal Gross { get; set; }
+    public double Gross { get; set; }
     public double Distance { get; set; }
 }
