+++ src/Core/Logistics.Application.Tenant.Contracts/Models/GrossesForIntervalDto.cs
@@ -19,6 +19,6 @@ public class GrossesForIntervalDto
         }
     }
     
-    public decimal TotalGross { get; private set; }
+    public double TotalGross { get; private set; }
     public double TotalDistance { get; private set; }
 }
