+++ src/Core/Logistics.Application.Tenant.Contracts/Models/LoadDto.cs
@@ -7,7 +7,7 @@ public class LoadDto
     public string? Name { get; set; }
     public string? SourceAddress { get; set; }
     public string? DestinationAddress { get; set; }
-    public decimal DeliveryCost { get; set; }
+    public double DeliveryCost { get; set; }
     public double Distance { get; set; }
     public DateTime DispatchedDate { get; set; }
     public DateTime? PickUpDate { get; set; }
